Update README.md
Question: should goal 2 "Goal: Coordinate English and Spanish versions of the lessons to keep them in synch." 
be something like : Goal: Coordinate English and Spanish versions of the lessons to to be updated in synch. Its related to our discussion of having flexibility of development, but we can versions that mature periodically. 

----
Comments, I think will be best to keep Carpentries with its english name